feat/15_blocked-executions: Case study cannot be deleted if case study has been executed

+++ apps/analyzer/views.py
@@ -288,7 +288,9 @@ def deleteCaseStudy(request):
     case_study_id = request.GET.get("id")
     cs = CaseStudy.objects.get(id=case_study_id)
     if request.user.id != cs.user.id:
-        raise Exception("This case study doesn't belong to the authenticated user")
+        raise Exception("This case study doesn't belong to the authenticated user");
+    if cs.executed != 0:
+        raise Exception("This case study cannot be deleted because it has already been excecuted");
     cs.delete()
     return HttpResponseRedirect(reverse("analyzer:casestudy_list"))
